requirements.txt
langchain 
langchain-groq
langchain-core
langchain-community
langchain-huggingface
langchain-chroma
sentence-transformers

streamlit
groq
pandas
numpy
chromadb
tqdm
python-dotenv

plotly

protobuf==3.20.3